Make entity collection builder abstract
